requirements.txt
fastapi>=0.75.0
uvicorn==0.40.0
python-multipart>=0.0.5
python-dotenv==1.1.1

# Reemplazo moderno compatible con Pydantic 2.x
fastapi-mail==1.4.1

# Whisper oficial (usa ffmpeg CLI, no PyAV)
openai-whisper==20250625

# Dependencias necesarias para Whisper
tqdm==4.66.4
tiktoken==0.7.0
regex==2024.11.6

# Pydantic 2.x stack
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pydantic-extra-types==2.10.5

email-validator>=1.1.3
cryptography==46.0.3

requests==2.32.3
urllib3==2.2.3
certifi==2024.12.14
charset-normalizer==3.4.0
idna==3.10

numpy==1.26.4
typing_extensions>=4.14.1
starlette>=0.40.0
click==8.1.8
h11==0.16.0

google-generativeai==0.7.2
google-ai-generativelanguage==0.6.6
google-api-core==2.29.0
googleapis-common-protos==1.72.0
grpcio==1.62.3
grpcio-status==1.62.3
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.1

yt-dlp==2026.1.29

# Development tools
flake8

# Local wheels
# ./wheels/openai_whisper-20250625-py3-none-any.whl
# ./wheels/onnxruntime_gpu-1.19.0-cp310-cp310-linux_aarch64.whl
# ./wheels/torch_cuda_jetpack-2.3.0-py3-none-any.whl
# ./wheels/torchaudio-2.3.0+952ea74-cp310-cp310-linux_aarch64.whl
# ./wheels/torchvision-0.18.0a0+6043bc2-cp310-cp310-linux_aarch64.whl
